requirements.txt
anyio==4.1.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cucim==23.10.0
cupy-cuda12x==12.3.0
cycler==0.12.1
Django==3.2.6
dm-tree==0.1.8
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fastrlock==0.8.2
filelock==3.13.1
filetype==1.2.0
fonttools==4.45.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.32.0
nvidia-dlprof==1.8.0
nvidia-dlprof-pytorch-nvtx==1.8.0
nvidia-dlprofviewer==1.8.0
nvidia-nccl-cu12==2.18.1
nvidia-nsys-cli==2021.3.2.12
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.9.26
torch==2.1.1
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
wandb==0.16.1
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
